refactor(webapp): move GoogleButton to components-v2/patterns
App-specific OAuth button doesn't belong in the v1 component tree.
Moves it to components-v2/patterns and adds a Storybook story.
