+++ packages/design-system/stories/v2-patterns/GoogleButton.stories.tsx
@@ -1,0 +1,18 @@
+import GoogleButton from '@/components-v2/patterns/GoogleButton';
+
+import type { Meta, StoryObj } from '@storybook/react-vite';
+
+const meta: Meta = {
+    title: 'Components v2/Patterns/GoogleButton',
+    parameters: { layout: 'padded' }
+};
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const Default: Story = {
+    render: () => <GoogleButton text="Sign in with Google" setServerErrorMessage={() => {}} />
+};
+
+export const SignUp: Story = {
+    render: () => <GoogleButton text="Sign up with Google" setServerErrorMessage={() => {}} />
+};
