+++ packages/webapp/src/components-v2/patterns/GoogleButton.tsx
@@ -1,4 +1,4 @@
-import { apiFetch } from '../../../../utils/api';
+import { apiFetch } from '../../utils/api';
 
 import type { PostManagedSignup } from '@nangohq/types';
